refactor: regroup Cangjie version service
