--- services/cangjie/cangjie-base.js
@@ -1,58 +1,0 @@
-import Joi from 'joi'
-import { BaseJsonlService, InvalidParameter, NotFound } from '../index.js'
-import { parseJsonl } from '../../core/base-service/jsonl.js'
-import { latest } from '../version.js'
-
-const indexEntrySchema = Joi.object({
-  organization: Joi.string(),
-  name: Joi.string().required(),
-  version: Joi.string().required(),
-  yanked: Joi.boolean().required(),
-  'index-version': Joi.string().required(),
-}).required()
-
-export class BaseCangjieService extends BaseJsonlService {
-  static defaultBadgeData = { label: 'cangjie' }
-
-  static getIndexPath({ moduleName }) {
-    if (moduleName.length < 3) {
-      throw new InvalidParameter({ prettyMessage: 'invalid module name' })
-    }
-
-    if (moduleName.length === 3) {
-      return `${moduleName.slice(0, 2)}/${moduleName[2]}/${moduleName}`
-    }
-
-    return `${moduleName.slice(0, 2)}/${moduleName.slice(2, 4)}/${moduleName}`
-  }
-
-  static getLatestVersion(entries) {
-    const version = latest(
-      entries.filter(({ yanked }) => !yanked).map(({ version }) => version),
-    )
-
-    if (version === undefined) {
-      throw new NotFound({ prettyMessage: 'no releases' })
-    }
-
-    return version
-  }
-
-  _parseJsonl(buffer) {
-    return parseJsonl(buffer, { prettyErrorMessage: 'invalid index entry' })
-  }
-
-  async fetch({ moduleName, organization }) {
-    const indexPath = this.constructor.getIndexPath({ moduleName })
-    const options = organization
-      ? { searchParams: { organization } }
-      : undefined
-
-    return this._requestJsonl({
-      schema: indexEntrySchema,
-      url: `https://pkg.cangjie-lang.cn/registry/index/${indexPath}`,
-      options,
-      prettyErrorMessage: 'invalid index entry',
-    })
-  }
-}
